refactor: decompose mod loading pipeline
